LayoutTests/media/hevc-codec-parameters-expected.txt

EXPECTED (internals.parseHEVCCodecParameters("bad-parameter") == 'null') OK
EXPECTED (internals.parseHEVCCodecParameters("hvc1") == 'null') OK
EXPECTED (internals.parseHEVCCodecParameters("hev1.1.6.L93") === '{ hev1, 0, 1, 1610612736, 0, [0,0,0,0,0,0], 93 }') OK
EXPECTED (internals.parseHEVCCodecParameters("hvc1.A4.41.H120") === '{ hvc1, 1, 4, 2181038080, 1, [0,0,0,0,0,0], 120 }') OK
EXPECTED (internals.parseHEVCCodecParameters("hev1.A4.41.H120.B0") === '{ hev1, 1, 4, 2181038080, 1, [176,0,0,0,0,0], 120 }') OK
EXPECTED (internals.parseHEVCCodecParameters("hvc1.B1.4.L63") === '{ hvc1, 2, 1, 536870912, 0, [0,0,0,0,0,0], 63 }') OK
EXPECTED (internals.parseHEVCCodecParameters("hev1.D1.4.L63") == 'null') OK
EXPECTED (internals.parseHEVCCodecParameters("hev1.B1.4.L68000") == 'null') OK
END OF TEST
